==> Data processing/01b - Load climate - ERA5 daily.Rmd <==
---
title: "Load climate"
author: "Abby Lewis"
date: "2022-10-10"
output: html_document
---

This script loads climate data from the ERA5 re-analysis data product and matches climate data to LakeIDs by lat/lon. Neither the raw ERA 5 data product nor the output of this file are small enough to store on GitHub, so data will need to be re-downloaded and re-processed for replication. 

To download ERA 5 daily temperature data: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form
  Product type: Reanalysis
  Variable: "2m temperature"
  Year: 1980-2022
  Month: all
  Day: all
  Time: 00:00
  Geographical Area: "Whole available region"
  Format: "NetCDF (experimental)"

Table of Contents:
- Step 1: Load packages
- Step 2: Load and synthesize data


Step 1: Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
Sys.setenv(TZ='UTC')
library(ncdf4)
library(fields)
library(tidyverse)
library(data.table)
library(lubridate)
```


Step 2: Load and synthesize data 
(note: this will take several minutes to run on a personal computer)
```{r}
#Read and format our lake metadata
lakes <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579") 
lakes_format <- lakes %>%
  dplyr::select(LakeID,Longitude_DD,Latitude_DD) %>%
  rename(Lon=Longitude_DD,Lat=Latitude_DD)%>%
  filter(!is.na(Lon),
         !is.na(Lat))

#Read NC temp
mynetcdf_t <- '../External data/era5_daily.nc' #Downloaded from here: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels-monthly-means?tab=form
#The downloaded data is too large to store on github
ncFile_t <- nc_open(mynetcdf_t)
Lon <- ncvar_get(ncFile_t,"longitude")
Lon <- ifelse(Lon > 180, -(360 - Lon), Lon)
Lat <- ncvar_get(ncFile_t,"latitude")
time <- ncvar_get(ncFile_t,"time")
fillvalue <- ncatt_get(ncFile_t, "t2m","missing_value") 

#Loop through lakes and add temp data based on lat/lon
combined_t_df <- data.frame(LakeID = rep(lakes_format$LakeID, each = length(time)),
                           Date = as.Date("1900-01-01")+hours(time),
                           Temp_C = as.numeric(NA),
                           Lon = as.numeric(NA),
                           Lat = as.numeric(NA)
                           )
for(i in 1:nrow(lakes_format)){
  print(lakes_format$LakeID[i])
  lon_id <- ifelse(min(abs(lakes_format[i,]$Lon-Lon)) <= 0.25, 
                   which.min(abs(lakes_format[i,]$Lon - Lon)), 
                   NA)
  lat_id <- ifelse(min(abs(lakes_format[i,]$Lat-Lat)) <= 0.25, 
                   which.min(abs(lakes_format[i,]$Lat - Lat)), 
                   NA)
  if(!is.na(lon_id) & !is.na(lat_id)){
    t_vector1 <- ncvar_get(ncFile_t,"t2m", start = c(lon_id, lat_id, 1, 1), count = c(1,1,1,-1))
    t_vector2 <- ncvar_get(ncFile_t,"t2m", start = c(lon_id, lat_id, 2, 1), count = c(1,1,1,-1))
    t_vector <- t_vector1
    t_vector[is.na(t_vector)] <- t_vector2[is.na(t_vector)]
    t_vector[t_vector == fillvalue$value] <- NA
    combined_t_df$Temp_C[combined_t_df$LakeID==lakes_format[i,]$LakeID]<- t_vector-273.15
    combined_t_df$Lon[combined_t_df$LakeID==lakes_format[i,]$LakeID]<- Lon[lon_id]
    combined_t_df$Lat[combined_t_df$LakeID==lakes_format[i,]$LakeID]<- Lat[lat_id]
  }
}

combined_t_df <- combined_t_df%>%
  mutate(Temp_C = trunc(Temp_C * 100000) / 100000)

write.csv(combined_t_df,"../Compiled data/historical_temp_output_era5_daily.csv", row.names = F)
```
