==> Data processing/03 - Summer avgs.Rmd <==
---
title: "Summer avgs"
author: "Abby Lewis"
date: "2022-11-22"
output: html_document
---

This file loads temp/DO data and calculates means during the late-summer period (July-Aug in the Northern Hemisphere). All code chunks must be run in order.

Table of contents:
Step 1: Load data and packages
Step 2: QAQC
Step 3: Filter to summer period and identify stratified lakes
Step 4: Calculate surface and bottom means
Step 5: Add buoyancy frequency during the late-summer period
Step 6: Plot
Step 7: Repeat steps 3, 4, and 5 for July 15-August 31
Step 8: Repeat steps 3, 4, and 5 for June/July/August

Step 1: Load data and packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)
source("thermo.depth.density.R")

## Load data and metadata from EDI
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
#Load interpolated DO data file, created by "02 - Temp and DO interpolation.Rmd"
do <- read.csv("../Compiled data/temp_o2_interpolated.csv")
p <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/0ece9d7b67cd49741ed7ee60192832e4")
```

Step 2: QAQC

```{r}
#Merge P, DO, and lake info (from lat long file)
full <- do %>%
  full_join(p) %>%
  left_join(lat_long, by = c("LakeID")) %>%
  mutate(Date = as.Date(Date),
         Date_unif = ifelse(Latitude_DD > 0,
                            Date,
                            Date + months(6)),
         Date_unif = as.Date(Date_unif)) %>%
  filter(is.na(MaximumDepth_m) | !MaximumDepth_m < 3,
         abs(Latitude_DD) >= 23.5)

full$Date_22 <- full$Date_unif
year(full$Date_22) <- 2022
```

Step 3: Filter to summer period and identify stratified lakes

```{r}
full_trimmed <- full %>%
  filter(Date_22 >= as.Date("2022-07-01") & Date_22 <= as.Date("2022-08-31")) %>%
  dplyr::select(-Date_22)

#Calculate thermocline depths
thermo_depths <- full_trimmed%>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(!is.na(Temp_C)) %>%
  ungroup() %>%
  group_by(Date_unif, LakeID) %>%
  dplyr::summarize(epi_depth = meta.depths(Temp_C,Depth_m,mixed.cutoff = 0)[1],
                   hypo_depth = meta.depths(Temp_C,Depth_m,mixed.cutoff = 0)[2],
                   max_depth = max(Depth_m),
                   #use custom density threshold function
                   thermo = thermo.depth.density(Temp_C, Depth_m, 
                                                 mixed.cutoff = 0.1, 
                                                 seasonal = F)) %>%
  mutate(Year = year(Date_unif),
         unstrat = as.numeric(is.na(thermo))) %>%
  group_by(Year,LakeID) %>%
  dplyr::summarize(epi_depth = mean(epi_depth, na.rm = T),
                   hypo_depth = mean(hypo_depth, na.rm = T),
                   count_unstrat = sum(unstrat),
                   n = n())

#How many are removed by filtering out lakes with 10% of profiles being unstratified? 77 lakes
thermo_depths%>%
  group_by(LakeID) %>%
  dplyr::mutate(count_unstrat_tot = sum(count_unstrat),
                n = sum(n)) %>%
  filter((count_unstrat_tot/n) >=0.1) %>%
  ungroup() %>%
  summarize(lakes = length(unique(LakeID)))

#Remove years with unstratified profiles and lakes where 10% of years have unstratified profiles
thermo_depths_sum <- thermo_depths%>%
  group_by(LakeID) %>%
  dplyr::mutate(count_unstrat_tot = sum(count_unstrat),
                n = sum(n)) %>%
  filter((count_unstrat_tot/n) <0.1,
         count_unstrat == 0) %>%
  group_by(LakeID) %>%
  dplyr::summarize(epi_sd = sd(epi_depth, na.rm = T),
            epi_depth = mean(epi_depth, na.rm = T),
            hypo_sd = sd(hypo_depth, na.rm = T),
            hypo_depth = mean(hypo_depth, na.rm = T))

#Export thermocline depths for reference
write_csv(thermo_depths_sum, "../Compiled data/stratified_lakes.csv")
```

Step 4: Calculate surface and bottom means

```{r}
surf_bot_daily <- full_trimmed %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T),
                   DO_mgL = mean(DO_mgL, na.rm = T),
                   DO_sat = mean(DO_sat, na.rm = T),
                   TP_ugL = mean(TP_ugL, na.rm = T),
                   DOC_mgL = mean(DOC_mgL, na.rm = T),
                   Chla_ugL = mean(Chla_ugL, na.rm = T)) %>%
  filter(!is.na(Temp_C) | !is.na(DO_mgL)) %>%
  group_by(LakeID, Date_unif) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL[Depth_m == 1], na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DO_sat_SURF = mean(DO_sat[Depth_m == 1], na.rm=T),
                   DO_sat_BOT = mean(DO_sat[Depth_m >= max(Depth_m)-1], na.rm=T),
                   Temp_C_SURF = mean(Temp_C[Depth_m == 1], na.rm=T),
                   Temp_C_BOT = mean(Temp_C[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL[Depth_m <= 1], na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL[Depth_m <= 1], na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL[Depth_m <= 1], na.rm=T)) 

#For summary stats
surf_bot_daily_all <- full %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T),
                   DO_mgL = mean(DO_mgL, na.rm = T),
                   DO_sat = mean(DO_sat, na.rm = T),
                   TP_ugL = mean(TP_ugL, na.rm = T),
                   DOC_mgL = mean(DOC_mgL, na.rm = T),
                   Chla_ugL = mean(Chla_ugL, na.rm = T)) %>%
  filter(!is.na(Temp_C) | !is.na(DO_mgL)) %>%
  group_by(LakeID, Date_unif) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL[Depth_m == 1], na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DO_sat_SURF = mean(DO_sat[Depth_m == 1], na.rm=T),
                   DO_sat_BOT = mean(DO_sat[Depth_m >= max(Depth_m)-1], na.rm=T),
                   Temp_C_SURF = mean(Temp_C[Depth_m == 1], na.rm=T),
                   Temp_C_BOT = mean(Temp_C[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL[Depth_m <= 1], na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL[Depth_m <= 1], na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL[Depth_m <= 1], na.rm=T)) 

write.csv(surf_bot_daily_all, "../Compiled data/surf_bot_daily_all.csv")
write.csv(surf_bot_daily, "../Compiled data/surf_bot_daily.csv")

#Summarize within each year
surf_bot <- surf_bot_daily %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL_SURF, na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL_BOT, na.rm=T),
                   DO_sat_SURF = mean(DO_sat_SURF, na.rm=T),
                   DO_sat_BOT = mean(DO_sat_BOT, na.rm=T),
                   Temp_C_SURF = mean(Temp_C_SURF, na.rm=T),
                   Temp_C_BOT = mean(Temp_C_BOT, na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL_SURF, na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL_SURF, na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL_SURF, na.rm=T))

#Calculate buoyancy frequency
buoyancy <- full_trimmed %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(!is.na(Temp_C)) %>%
  ungroup() %>%
  group_by(Date_unif, LakeID) %>%
  dplyr::summarize(buoyancy_freq = max(buoyancy.freq(Temp_C,Depth_m),na.rm=T)) %>%
  mutate(Year=year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(buoyancy_freq = mean(buoyancy_freq,na.rm=T))
```

Step 5: Add buoyancy frequency during the late-summer period

```{r}
summer_avgs <- surf_bot %>%
  left_join(buoyancy) 

#All done!
write.csv(summer_avgs, "../Compiled data/summer_averages_wi.csv", row.names = F)
```

Step 6: Plot 

```{r}
jpeg("../Figures/Data_hist.jpg", width = 3, height = 3, units = "in", res = 300)
all <- full %>%
  group_by(LakeID) %>%
  filter(length(unique(year(Date_unif))) >= 15) %>%
  filter(Depth_m == 1) %>%
  group_by(Date_22, LakeID) %>%
  summarize(n = length(!is.na(Temp_C))) %>%
  filter(n > 0) %>%
  mutate(month = as.numeric(month(Date_22)),
         focal = ifelse(month %in% c(7,8),
                        "Summer",
                        "All"),
         title = "Annual") %>%
  ggplot(aes(x = month, fill = focal)) +
  geom_histogram(bins = 12) +
  scale_x_continuous(breaks = seq(1, 12, by = 2), labels = month.abb[seq(1, 12, by = 2)])+
  ylab("Total number of sampling events") +
  scale_fill_manual(values = c("grey50", "grey20"))+
  theme_bw() +
  facet_wrap(~title)+
  theme(axis.title.x = element_blank(),
        legend.position = "none")
all
dev.off()

for_summer_plot <- full %>%
  group_by(LakeID) %>%
  filter(length(unique(year(Date_unif))) >= 15) %>%
  ungroup() %>%
  mutate(Date_unif = as.Date(Date_unif)) %>%
  filter(!is.na(Temp_C),
         Depth_m ==1,
         month(Date_unif) %in% c(6:9)) %>%
  mutate(focal = ifelse(month(Date_unif) %in% c(7,8),
                        "Summer",
                        "All")) 

summer <- for_summer_plot %>%
  mutate(title = "Summer") %>%
  ggplot(aes(x = Date_22, fill = focal)) +
  geom_histogram(bins = length(unique(for_summer_plot$Date_22))) +
  ylab("Total number of sampling events") +
  scale_fill_manual(values = c("grey50", "grey20"))+
  scale_x_date(breaks = as.Date(c("2022-07-01", "2022-08-01", "2022-09-01")),
               labels = c(month.abb[7:9]))+
  theme_bw() +
  facet_wrap(~title)+
  theme(axis.title.x = element_blank(),
        legend.position = "none",
        axis.title.y = element_blank())+
  coord_cartesian(xlim = as.Date(c("2022-07-01", "2022-08-31")))

for_n_dates <- for_summer_plot %>%
  filter(focal == "Summer") %>%
  mutate(year = year(Date_unif)) %>%
  group_by(year, LakeID) %>%
  summarize(n = length(unique(Date_unif))) %>%
  group_by(LakeID) %>%
  summarize(n = mean(n))

n_dates <- for_n_dates %>%
  ggplot(aes(x = n))+
  xlab("Sampling dates per summer")+
  geom_vline(xintercept = median(for_n_dates$n))+
  geom_text(aes(x = x + 0.5),
            y = 0.35,
            label = paste0("Median: ", round(median(for_n_dates$n),0), " sampling dates"),
            hjust = 0,
            data = data.frame(x = median(for_n_dates$n)))+
  geom_density()+
  theme_bw()

jpeg("../figures/date_dist.jpg", width = 6, height = 4, units = "in", res = 300)
p <- ggpubr::ggarrange(all + theme(axis.title.y = element_blank()), 
                       ggpubr::ggarrange(summer, n_dates, labels = c("b", "c")), 
                       labels = c("a", NA), nrow = 2)
ggpubr::annotate_figure(p, left = grid::textGrob("Total number of sampling events", 
                                         rot = 90, vjust = 1, 
                                         gp = grid::gpar(cex = .9)))
dev.off()

surf_bot_table <- surf_bot_daily %>%
  pivot_longer(c(DO_mgL_SURF, DO_mgL_BOT, DO_sat_SURF, DO_sat_BOT, Temp_C_SURF, 
                 Temp_C_BOT, DOC_mgL_SURF, Chla_ugL_SURF, TP_ugL_SURF)) %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(LakeID) %>%
  filter(length(unique(Year)) >= 15) %>%
  group_by(LakeID, name, Year) %>%
  filter(!is.na(value)) %>%
  summarize(n = n()) %>%
  group_by(LakeID, name) %>%
  summarize(mean_n = mean(n)) %>%
  group_by(name) %>%
  summarize(median_n = median(mean_n, na.rm = T))
```

Step 6: Repeat steps 3, 4, and 5 for July 15-August 31

```{r}
full_trimmed <- full %>%
  filter(Date_22>=as.Date("2022-07-01")&Date_22<=as.Date("2022-08-31")) %>%
  dplyr::select(-Date_22)

surf_bot_daily <- full_trimmed %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T),
                   DO_mgL = mean(DO_mgL, na.rm = T),
                   DO_sat = mean(DO_sat, na.rm = T),
                   TP_ugL = mean(TP_ugL, na.rm = T),
                   DOC_mgL = mean(DOC_mgL, na.rm = T),
                   Chla_ugL = mean(Chla_ugL, na.rm = T)) %>%
  filter(!is.na(Temp_C) | !is.na(DO_mgL)) %>%
  group_by(LakeID, Date_unif) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL[Depth_m == 1], na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DO_sat_SURF = mean(DO_sat[Depth_m == 1], na.rm=T),
                   DO_sat_BOT = mean(DO_sat[Depth_m >= max(Depth_m)-1], na.rm=T),
                   Temp_C_SURF = mean(Temp_C[Depth_m == 1], na.rm=T),
                   Temp_C_BOT = mean(Temp_C[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL[Depth_m <= 1], na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL[Depth_m <= 1], na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL[Depth_m <= 1], na.rm=T)) 

#Summarize within each year
surf_bot <- surf_bot_daily %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL_SURF, na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL_BOT, na.rm=T),
                   DO_sat_SURF = mean(DO_sat_SURF, na.rm=T),
                   DO_sat_BOT = mean(DO_sat_BOT, na.rm=T),
                   Temp_C_SURF = mean(Temp_C_SURF, na.rm=T),
                   Temp_C_BOT = mean(Temp_C_BOT, na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL_SURF, na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL_SURF, na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL_SURF, na.rm=T))

#Calculate buoyancy frequency
buoyancy <- full_trimmed %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(!is.na(Temp_C)) %>%
  ungroup() %>%
  group_by(Date_unif, LakeID) %>%
  dplyr::summarize(buoyancy_freq = max(buoyancy.freq(Temp_C,Depth_m),na.rm=T)) %>%
  mutate(Year=year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(buoyancy_freq = mean(buoyancy_freq,na.rm=T))

summer_avgs <- surf_bot %>%
  left_join(buoyancy) 

#All done!
write.csv(summer_avgs, "../Compiled data/summer_averages_wi_15th_to_31st.csv", row.names = F)
```

Step 7: Repeat steps 3, 4, and 5 for June/July/August

```{r}
full_date <- full %>%
  filter(Date_22>=as.Date("2022-07-01")&Date_22<=as.Date("2022-08-31")) %>%
  dplyr::select(-Date_22)

surf_bot_daily <- full_trimmed %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T),
                   DO_mgL = mean(DO_mgL, na.rm = T),
                   DO_sat = mean(DO_sat, na.rm = T),
                   TP_ugL = mean(TP_ugL, na.rm = T),
                   DOC_mgL = mean(DOC_mgL, na.rm = T),
                   Chla_ugL = mean(Chla_ugL, na.rm = T)) %>%
  filter(!is.na(Temp_C) | !is.na(DO_mgL)) %>%
  group_by(LakeID, Date_unif) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL[Depth_m == 1], na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DO_sat_SURF = mean(DO_sat[Depth_m == 1], na.rm=T),
                   DO_sat_BOT = mean(DO_sat[Depth_m >= max(Depth_m)-1], na.rm=T),
                   Temp_C_SURF = mean(Temp_C[Depth_m == 1], na.rm=T),
                   Temp_C_BOT = mean(Temp_C[Depth_m >= max(Depth_m)-1], na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL[Depth_m <= 1], na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL[Depth_m <= 1], na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL[Depth_m <= 1], na.rm=T)) 

#Summarize within each year
surf_bot <- surf_bot_daily %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(DO_mgL_SURF = mean(DO_mgL_SURF, na.rm=T),
                   DO_mgL_BOT = mean(DO_mgL_BOT, na.rm=T),
                   DO_sat_SURF = mean(DO_sat_SURF, na.rm=T),
                   DO_sat_BOT = mean(DO_sat_BOT, na.rm=T),
                   Temp_C_SURF = mean(Temp_C_SURF, na.rm=T),
                   Temp_C_BOT = mean(Temp_C_BOT, na.rm=T),
                   DOC_mgL_SURF = mean(DOC_mgL_SURF, na.rm=T),
                   Chla_ugL_SURF = mean(Chla_ugL_SURF, na.rm=T),
                   TP_ugL_SURF = mean(TP_ugL_SURF, na.rm=T))

#Calculate buoyancy frequency
buoyancy <- full_trimmed %>%
  group_by(Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(!is.na(Temp_C)) %>%
  ungroup() %>%
  group_by(Date_unif, LakeID) %>%
  dplyr::summarize(buoyancy_freq = max(buoyancy.freq(Temp_C,Depth_m),na.rm=T)) %>%
  mutate(Year=year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(buoyancy_freq = mean(buoyancy_freq,na.rm=T))

summer_avgs <- surf_bot %>%
  left_join(buoyancy) 

#All done!
write.csv(summer_avgs, "../Compiled data/summer_averages_wi_JJA.csv", row.names = F)
```
