==> Data processing/04 - Stratified avgs.Rmd <==
---
title: "Stratified averages"
author: "Abby Lewis"
date: "2024-01-25"
output: html_document
---

This file loads temp/DO data from the whole year (not just summer), calculates metrics of stratification, exports stratification onset dates, and exports a compiled file with data during the entire stratified period (used to calculate oxygen demand). All code chunks must be run in order.

Table of contents:
Step 1: Load data and packages
Step 2: Filter to stratified period
Step 3: Calculate stratification onset to export 
Step 4: Calculate metalimetic boundaries and average values within each layer

Step 1: Load data and packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)
source("thermo.depth.density.R")

## Load data and metadata from EDI
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
#Load interpolated DO data file, created by "02 - Temp and DO interpolation.Rmd"
do <- read.csv("../Compiled data/temp_o2_interpolated.csv")
#Merge data and lake info (from lat long file)
full <- do %>%
  full_join(lat_long, by = c("LakeID")) %>%
  mutate(Date = as.Date(Date),
         Date_unif = ifelse(Latitude_DD > 0,
                            Date,
                            Date + months(6)),
         Date_unif = as.Date(Date_unif)) %>%
  filter(!is.na(MaximumDepth_m) & MaximumDepth_m >= 3,
         abs(Latitude_DD) >= 23.5)
```

Step 2: Filter to stratified period

```{r}
#Create a Date_22 column, so that I can filter to only the end of summer
full$Date_22 <- full$Date_unif
year(full$Date_22) <-2022

#There are a few rows with missing dates because these lakes are in the lat long database and not do
full %>%
  filter(is.na(as.Date(Date_unif)))
#Remove these NAs
full <- full %>%
  filter(!is.na(Date_unif),
         !is.na(Latitude_DD))

###
#Determine start and end of stratification, then thermo depths
###

full_format <- full %>%
  group_by(Date, Date_unif, LakeID, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(!is.na(Temp_C)) %>%
  ungroup() %>%
  group_by(Date, Date_unif, LakeID) %>%
  group_by(Date_unif, LakeID) %>%
  dplyr::summarize(epi_depth = meta.depths(Temp_C,Depth_m,mixed.cutoff = 0)[1],
                   hypo_depth = meta.depths(Temp_C,Depth_m,mixed.cutoff = 0)[2],
                   max_depth = max(Depth_m),
                   #use custom density threshold function
                   thermo = thermo.depth.density(Temp_C, Depth_m, 
                                                 mixed.cutoff = 0.1, 
                                                 seasonal = F)) %>%
  mutate(Year = year(Date_unif),
         unstrat = as.numeric(is.na(thermo)))

#Start of stratification
date_start <- full_format %>%
  group_by(Year, LakeID) %>%
  filter(unstrat == 1,
         month(Date_unif) <= 7) %>%
  filter(Date_unif == max(Date_unif)) %>%
  dplyr::select(Date_unif, Year, LakeID) %>%
  rename(Date_start = Date_unif)

#End of stratification
date_end <- full_format %>%
  group_by(Year,LakeID) %>%
  filter(unstrat == 1,
         month(Date_unif) > 7) %>%
  filter(Date_unif==min(Date_unif)) %>%
  dplyr::select(Date_unif, Year, LakeID) %>%
  rename(Date_end = Date_unif)

#Stratification duration
strat_dur <- date_start%>%
  full_join(date_end)

#Filter to stratified period
full_trimmed <- full %>%
  mutate(Year = year(Date_unif)) %>%
  full_join(strat_dur) %>%
  filter(is.na(Date_start) | Date_unif>Date_start,
         is.na(Date_end) | Date_unif<Date_end,
         month(Date_unif) > 3,
         month(Date_unif) < 11)
```

Step 3: Calculate stratification onset to export 
(slightly different calculation to be more strict about precisely calculating the onset of stratification)

```{r}
#Start of stratification
date_start <- full_format%>%
  group_by(Year, LakeID) %>%
  filter(unstrat == 1,
         month(Date_unif) <= 6) %>%
  filter(Date_unif==max(Date_unif)) %>%
  dplyr::select(Date_unif, Year, LakeID) %>%
  rename(Date_start = Date_unif)

#Calculate stratification onset dates if we can quantify this within 7 days
export <- full_format%>%
  left_join(date_start) %>%
  filter(unstrat == 0,
         Date_unif > Date_start) %>%
  group_by(Year, LakeID) %>%
  filter(Date_unif == min(Date_unif)) %>%
  dplyr::select(Date_unif, Year, LakeID, Date_start) %>%
  rename(Date_next = Date_unif) %>%
  filter(Date_next - Date_start < 7) %>%
  mutate(Onset = median(yday(c(Date_start, Date_next))))

#Limit to lakes that are included in the rest of the analysis
lakes <- read.csv("../Compiled data/stratified_lakes.csv")

#Calculate median at each lake
export2 <- export%>%
  filter(LakeID %in% lakes$LakeID) %>%
  group_by(LakeID) %>%
  summarize(Onset = median(Onset, na.rm = T))

#Export
write.csv(export2, "../Compiled data/stratification_onset.csv", row.names = F)
```

Step 4: Calculate metalimetic boundaries and average values within each layer

```{r}
#Thermocline depths during stratified period
thermo_depths <- full_format%>%
  full_join(strat_dur) %>%
  filter(is.na(Date_start) | Date_unif > Date_start,
         is.na(Date_end) | Date_unif < Date_end) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(epi_depth = mean(epi_depth, na.rm = T),
            hypo_depth = mean(hypo_depth, na.rm = T),
            count_unstrat = sum(unstrat),
            n= n())

thermo_depths_sum <- thermo_depths%>%
  group_by(LakeID) %>%
  dplyr::mutate(count_unstrat_tot = sum(count_unstrat),
                n = sum(n)) %>%
  filter((count_unstrat_tot/n) < 0.1,
         count_unstrat == 0) %>%
  group_by(LakeID,Year) %>%
  dplyr::summarize(epi_sd = sd(epi_depth, na.rm = T),
            epi_depth = mean(epi_depth, na.rm = T),
            hypo_sd = sd(hypo_depth, na.rm = T),
            hypo_depth = mean(hypo_depth, na.rm = T))

full_with_thermo <- full_trimmed%>%
  full_join(thermo_depths_sum) %>%
  filter(!is.na(epi_depth),
         !is.na(hypo_depth))

write.csv(full_with_thermo,"../Compiled data/Stratified_period_data_with_thermo.csv",row.names = F)
```
