==> Data processing/05 - VW DO Demand - based on strat dur.Rmd <==
---
title: "ASL DO Demand calcs"
author: "Abby Lewis"
date: "2022-08-29"
output: html_document
---

This file calculates volume-weighted hypolimnetic oxygen demand during the stratified period at each lake. 

Table of contents:
Step 1: Load data and packages
Step 2: Calculate the rate of change in volume-weighted oxygen concentrations during the stratified period at each lake (VHOD)


Step 1: Load data and packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)

# Load data and stratification boundaries. This file is created by "03 - Stratified avgs.Rmd"
full_with_thermo <- read.csv("../Compiled data/Stratified_period_data_with_thermo.csv") %>%
  mutate(Date_original = as.Date(Date),
         Date_unif = as.Date(Date_unif))
# Load lake metadata
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
```

Step 2: Calculate the rate of change in volume-weighted oxygen concentrations during the stratified period at each lake (VHOD)

```{r}
# Prepare bathymetry
lake_bats <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/ba733454c29a5f026efed05a6e6ef75b") %>%
  filter(Area_m2>0) %>%
  group_by(LakeID) %>%
  mutate(methods_n = length(unique(Method))) %>%
  filter(methods_n==1|Method=="data") %>% #when possible use data rather than model
  ungroup()

# Check bathymetric data availability
unique(lat_long$LakeID)[!unique(lat_long$LakeID) %in% unique(lake_bats$LakeID)] #missing full bathymetry for 37 lakes

# Total n to start: 739
length(unique(full_with_thermo$LakeID))

# Calculate volume-weighted oxygen concentrations
vw_do <- full_with_thermo %>%
  full_join(lake_bats, by = c("LakeID","Depth_m")) %>%
  filter(!is.na(IntervalVolume_m3)) %>%
  filter(Depth_m > hypo_depth) %>%
  mutate(DO_mass = DO_mgL * IntervalVolume_m3,
         Temp_total = Temp_C * IntervalVolume_m3,
         Year = year(Date_unif)) %>%
  group_by(Date_unif, LakeID) %>%
  dplyr::summarize(DO_tot = sum(DO_mass),#sum across all hypolimnetic depths
                   vol_tot = sum(IntervalVolume_m3),
                   DO_mgL_vol = DO_tot / vol_tot,#Divide by hypolimnetic volume
                   DO_mgL_area = DO_tot / max(Area_m2),
                   Temp_C = sum(Temp_total) / vol_tot) %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(LakeID, Year) %>%
  filter(!is.na(DO_mgL_vol)) %>%
  arrange(LakeID, Date_unif) %>%
  #Identify the lowest DO value for a given year
  mutate(low_point = ifelse(sum(DO_mgL_vol < 1)==0,
                            Date_unif[which.min(DO_mgL_vol)],
                            first(Date_unif[DO_mgL_vol < 1]))) %>%
  filter(is.na(low_point) | Date_unif <= low_point) %>% #Remove days after the lowest DO value
  ungroup()

length(unique(vw_do$LakeID)) #689 lakes left

# Calculate the rate of change in volume-weighted concentrations using lm
vw_do_demand <- vw_do %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(Year, LakeID) %>%
  dplyr::summarize(n = n(),
            DO_demand_mgLd = -lm(DO_mgL_vol~Date_unif)$coefficients[2],
            AHOD_mgLd = -lm(DO_mgL_area~Date_unif)$coefficients[2],
            p_vol = summary(lm(DO_mgL_vol~Date_unif))$coefficients[8],
            p_area = summary(lm(DO_mgL_area~Date_unif))$coefficients[8],
            r2_vol = summary(lm(DO_mgL_vol~Date_unif))$r.squared,
            r2_area = summary(lm(DO_mgL_area~Date_unif))$r.squared,
            Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(n >= 3) #Need at least 3 points
length(unique(vw_do_demand$LakeID[!is.na(vw_do_demand$DO_demand_mgLd)])) #466

# How many rows get removed by filtering so R2>0.5? 246
removed_by_r2 <- vw_do_demand %>%
  filter(r2_vol > .5)
nrow(removed_by_r2[!is.na(removed_by_r2$DO_demand_mgLd),]) #5447
nrow(vw_do_demand[!is.na(vw_do_demand$DO_demand_mgLd),]) #5736

# How many rows get removed by filtering so oxygen demand is positive? 0
removed_by_pos <- removed_by_r2 %>%
  filter(DO_demand_mgLd>0)
nrow(removed_by_pos[!is.na(removed_by_pos$DO_demand_mgLd),])

# Finish QAQC
good_hod <- vw_do_demand %>%
  group_by(LakeID) %>%
  filter(!is.na(DO_demand_mgLd),
         r2_vol>0.5
         ) %>%
  dplyr::summarize(n_neg = sum(DO_demand_mgLd<0),
                   n = n(),
                   pct = n_neg/n*100) %>%
  filter(pct <= 10)
length(unique(good_hod$LakeID)) #459

vw_do_demand_qaqc <- removed_by_pos %>%
  filter(LakeID %in% good_hod$LakeID)

nrow(vw_do_demand_qaqc[!is.na(vw_do_demand_qaqc$DO_demand_mgLd),]) #5447
length(unique(vw_do_demand_qaqc$LakeID[!is.na(vw_do_demand_qaqc$DO_demand_mgLd)])) #459

# Success!
write.csv(vw_do_demand_qaqc, "../Compiled data/VW oxygen demand points.csv", row.names = F)
```
