==> Data processing/06 - Compile data.Rmd <==
---
title: "Compile data"
author: "Abby Lewis"
date: "2022-09-21"
output: html_document
---

This document synthesizes all of the calculated data saved up to this point and outputs a file for later analyses. 

Table of contents:
Step 1: Load packages
Step 2: Compile all in-lake data


Step 1: Load packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(viridis)
library(lubridate)
```

Step 2: Compile all in-lake data

```{r}
do_demand_points <- read.csv("../Compiled data/VW oxygen demand points.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")

summer_avgs = read.csv("../Compiled data/summer_averages_wi.csv")
summer_avgs_wide <- summer_avgs%>%
  left_join(do_demand_points%>%
              dplyr::select(DO_demand_mgLd, LakeID, Year))

length(unique(summer_avgs_wide$LakeID))

with_temp_3m <- summer_avgs_wide %>%
  left_join(lat_long, by = c("LakeID"))%>%
  filter(MaximumDepth_m>=3,
         !LakeID=="387") #Hypolimnetic oxygenation

write.csv(with_temp_3m,"../Compiled data/All_data_annual_3m.csv", row.names =F)
```

Step 3: Compile data with summer defined as July 15-Aug 31

```{r}
summer_avgs = read.csv("../Compiled data/summer_averages_wi_15th_to_31st.csv") %>%
  unique()

with_temp <- summer_avgs_wide %>%
  left_join(lat_long, by = c("LakeID"))%>%
  filter(MaximumDepth_m>3,
         !LakeID=="387") #Hypolimnetic oxygenation

write.csv(with_temp,"../Compiled data/All_data_annual_15th_to_31st.csv", row.names = F)
```

Step 2: Compile data with summer defined as June, July, and August

```{r}
summer_avgs = read.csv("../Compiled data/summer_averages_wi_JJA.csv") %>%
  unique()

with_temp <- summer_avgs_wide %>%
  left_join(lat_long, by = c("LakeID"))%>%
  filter(MaximumDepth_m>3,
         !LakeID=="387") #Hypolimnetic oxygenation

write.csv(with_temp,"../Compiled data/All_data_annual_JJA.csv", row.names =F)
```
