==> Data processing/07 - ISIMIP_v3.Rmd <==
---
title: "ISIMIP analysis"
author: "Abby Lewis"
date: "2024-01-19"
output: html_document
---

This file loads ISIMIP modeled water temperature and formats for later analysis. Note that interpolation takes a long time (~an hour on my laptop)

Table of contents:

- Step 1: Load packages and lake metadata
- Step 2: Compile to epi/hypo means
- Step 3: Load air temp data

Step 1: Load packages and lake metadata

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ncdf4)
library(tidyverse)
library(lubridate)
library(pracma)
library(rLakeAnalyzer)
library(furrr)
source("../Data processing/thermo.depth.density.R")

lakes <- read.csv("lakes_for_isimip.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")
```

Step 1: Load all files
```{r}
hylak_ref <- nc_open("../External data/hydrolakes_id.nc")
files <- list.files("../External data/files_gotm/")

load_gotm_file <- function(file){
  test <- nc_open(paste0("../External data/files_gotm/",file))
  isimip_id <- sub("20crv3-era5_historical_obsclim_gotm_", "", 
                sub("_daily_1901_2021.nc", "", file))
  Lon <- ncvar_get(test,"lon")
  Lon <- ifelse(Lon > 180, -(360 - Lon), Lon)
  Lat <- ncvar_get(test,"lat")
  time <- ncvar_get(test,"time")
  date <- as.Date("1881-01-01")+seconds(time)
  depth <- ncvar_get(test,"z")
  t_vector <- ncvar_get(test,"temp")
  fillvalue <- test[["var"]][["temp"]][["missval"]]
  t_vector[t_vector == fillvalue] <- NA
  hylak_id_df <- ncvar_get(hylak_ref,"id")
  hylak_lon <- ncvar_get(hylak_ref,"lon")
  hylak_lat <- ncvar_get(hylak_ref,"lat")
  hylak_id <- hylak_id_df[which(hylak_lon == c(Lon)), 
                          which(hylak_lat == c(Lat))]
  
  #Loop through lakes and add temp data based on lat/lon
  combined_t_df <- data.frame(hylak_id = rep(hylak_id, dim(depth)[1]*dim(depth)[2]),
                             Date = rep(date, each = dim(depth)[1]),
                             Depth_m = -c(depth), #I want positive depths
                             Temp_C = c(t_vector),
                             Lon = Lon,
                             Lat = Lat
                             )
  return(combined_t_df)
}

all_gotm <- map(files, load_gotm_file) %>%
  bind_rows() %>%
  filter(year(Date) >= 1980)
```

Step 2: Interpolate

```{r}
#Prep for interpolation
to_profiles_cleaning1 <- all_gotm %>%
  filter(!is.na(Temp_C)) %>%
  mutate(Date = as.Date(Date)) %>%
  group_by(hylak_id, Date) %>% 
  filter(length(unique(Depth_m)) > 2, #less than three measurement depths
         )%>%
  mutate(crit = min(Depth_m)) %>% 
  ungroup() %>%
  filter(crit <= 3) #remove profiles where the lowest value is larger than 3

#take care of all the non-zeros
nozeros <- to_profiles_cleaning1 %>% 
  group_by(hylak_id, Date) %>%
  mutate(crit=min(Depth_m)) %>% 
  ungroup() %>%
  filter(crit != 0)

#change anything Depth <0.5 to 0 (remove them first)
nozeros_case1 <- nozeros %>% 
  filter(crit <= 0.5) %>%
  mutate(Depth_m = ifelse(Depth_m == crit, 0, Depth_m))

#add additional 0m measurement for depths between 0.5 and 3
nozeros_case2_zeros <- nozeros %>% 
  filter(crit > 0.5) %>% 
  filter(crit <= 3) %>%
  ungroup() %>%
  filter(Depth_m == crit)%>%#only take the lowest value
  mutate(Depth_m=0) %>% unique()

nozeros_case2 = nozeros %>% 
  filter(crit > 0.5) %>% 
  filter(crit <= 3) %>%
  full_join(nozeros_case2_zeros)

to_profiles_cleaning2 <- to_profiles_cleaning1 %>% 
  filter(crit == 0) %>%
  ungroup() %>%
  rbind(nozeros_case1) %>%
  rbind(nozeros_case2) %>% 
  mutate(Date = as.Date(Date)) %>%
  arrange(hylak_id, Date, Depth_m)  %>%
  group_by(hylak_id, Date) %>% 
  mutate(somany=n_distinct(Depth_m)) %>% 
  ungroup() %>% 
  filter(somany>=2) %>% #less than three measurement depths
  dplyr::select(-somany)

interp_t <- function(lake, to_profiles_cleaning2){
  alldates<-to_profiles_cleaning2 %>% 
    dplyr::select(hylak_id, Date, Depth_m, Temp_C) %>%
    filter(hylak_id == lake) %>%
    dplyr::select(Date) %>% 
    unique() %>% 
    unlist() %>% 
    as.vector()
  
  lake_data <- to_profiles_cleaning2 %>%
    filter(hylak_id == lake)
  
  interpolatedvalues.do <- furrr::future_map(alldates, run_all_dates_t, lake_data,lake) |> list_rbind()
  
  return(interpolatedvalues.do)
}

run_all_dates_t <- function(alldates, lake_data, lake){
  profile <- lake_data %>% 
    filter(Date == alldates)
  
  interpolated.depths <- seq(from=0, to= max(profile$Depth_m))
  
  datamatrix_<- data.frame(
    depths=profile$Depth_m,
    values=profile$Temp_C) %>%
    group_by(depths) %>% 
    dplyr::summarize(valuemean=mean(values)) %>% 
    ungroup() %>% 
    as.matrix()
  
  interpolated.dos <- pchip(xi= datamatrix_[,1], 
                            yi=datamatrix_[,2],
                            x= interpolated.depths)
  
  output = data.frame(hylak_id = lake,
                      Date = alldates,
                      Depth_m = interpolated.depths,
                      Temp_C = interpolated.dos)
  
  return(output)
}

plan(multisession)

laken <- to_profiles_cleaning2$hylak_id %>% 
  unique()

#NOTE: this step is highly computationally intensive
interpolatedvalues.t <- furrr::future_map(laken,interp_t,to_profiles_cleaning2) |> list_rbind()
```

Step 2: Compile to epi/hypo means

```{r}
isimip_summer <- interpolatedvalues.t %>%
  mutate(hylak_id = as.numeric(hylak_id)) %>%
  left_join(lat_long) %>%
  mutate(Date = as.Date(Date, origin = "1970-01-01"),
         Date_unif = ifelse(Latitude_DD < 0, 
                            Date + months(6),
                            Date),
         Date_unif = as.Date(Date_unif)) %>%
  filter(((Latitude_DD > 0) & 
            yday(Date_unif) >= yday("2022-07-01") & 
            yday(Date_unif) <= yday("2022-08-31")))

surf_bot_daily <- isimip_summer %>%
  group_by(Date_unif, LakeID, hylak_id, Depth_m) %>%
  dplyr::summarize(Temp_C = mean(Temp_C, na.rm = T)) %>%
  filter(!is.na(Temp_C)) %>%
  group_by(LakeID, Date_unif, hylak_id) %>%
  dplyr::summarize(Temp_C_SURF = mean(Temp_C[Depth_m == 1], na.rm=T),
                   Temp_C_BOT = mean(Temp_C[Depth_m >= max(Depth_m)-1], na.rm=T)) 

#Summarize within each year
surf_bot <- surf_bot_daily %>%
  mutate(Year = year(Date_unif)) %>%
  group_by(Year, LakeID, hylak_id) %>%
  dplyr::summarize(Temp_C_SURF = mean(Temp_C_SURF, na.rm=T),
                   Temp_C_BOT = mean(Temp_C_BOT, na.rm=T))

unstrat <- interpolatedvalues.t %>%
  mutate(hylak_id = as.numeric(hylak_id)) %>%
  left_join(lat_long) %>%
  mutate(Date = as.Date(Date, origin = "1970-01-01"),
         Date_unif = ifelse(Latitude_DD < 0, 
                            Date + months(6),
                            Date),
         Date_unif = as.Date(Date_unif)) %>%
  filter(month(Date_unif) < 7) %>%
  group_by(Date_unif, LakeID, hylak_id) %>%
  dplyr::summarize(max_depth = max(Depth_m),
            thermo = thermo.depth.density(Temp_C, Depth_m, mixed.cutoff = 0.1, seasonal = F))%>% 
  mutate(Year= year(Date_unif),
         unstrat = as.numeric(is.na(thermo)))%>%
  filter(unstrat == 1,
         !max_depth > 1E19)

strat_onset <- unstrat %>%
  group_by(LakeID, Year) %>%
  filter(Date_unif == max(Date_unif)) %>%
  group_by(LakeID, hylak_id) %>%
  summarize(strat_onset = median(yday(Date_unif), na.rm = T))

write.csv(strat_onset, "../Compiled data/ISIMIP_strat_onset_v3.csv", row.names = F)

strat_onset_yearly <- unstrat %>%
  group_by(LakeID, Year, hylak_id) %>%
  filter(Date_unif == max(Date_unif))

write.csv(strat_onset_yearly, "../Compiled data/ISIMIP_strat_onset_v3_yearly.csv", row.names = F)

write.csv(surf_bot, "../Compiled data/ISIMIP_summer_avgs_v3 - surf bot.csv", row.names = F)
```

Step 3: Load air temp. data

```{r}
hylak_ref <- nc_open("../External data/hydrolakes_id.nc")
files <- list.files("../External data/lake_tas/")

load_tas_file <- function(file){
  test <- read_tsv(paste0("../External data/lake_tas/",file),
                   col_names = c("Date", "Time", "Temp_C"),
                   show_col_types = FALSE)
  isimip_id <- sub("20crv3-era5_historical_obsclim_lake_", "", 
                sub("_daily_1901_2021.csv", "", file))
  
  #Find hydrolakes ID
  nc_file <- nc_open(paste0("../External data/files_gotm/20crv3-era5_historical_obsclim_gotm_", 
                            isimip_id, "_daily_1901_2021.nc"))
  Lon <- ncvar_get(nc_file,"lon")
  Lon <- ifelse(Lon > 180, -(360 - Lon), Lon)
  Lat <- ncvar_get(nc_file,"lat")
  hylak_id_df <- ncvar_get(hylak_ref,"id")
  hylak_lon <- ncvar_get(hylak_ref,"lon")
  hylak_lat <- ncvar_get(hylak_ref,"lat")
  hylak_id <- hylak_id_df[which(hylak_lon == c(Lon)), 
                          which(hylak_lat == c(Lat))]
  
  #Add temp data based on lat/lon
  combined_t_df <- data.frame(hylak_id = rep(hylak_id, nrow(test)),
                             Date = test$Date,
                             Temp_C = test$Temp_C - 273.15,
                             Lon = Lon,
                             Lat = Lat
                             )
  return(combined_t_df)
}

all_tas <- map(files, load_tas_file) %>%
  bind_rows() %>%
  filter(year(Date) >= 1980)

write.csv(all_tas, "../Compiled data/ISIMIP_ERA5_temp_1981_2005.csv", row.names = F)
```
